# [Performance] (Java exec beats Ruby exec inside!) Faster Java Execution Using CodeGen/Janino