--- a/logstash-core/lib/logstash-core_jars.rb
+++ b/logstash-core/lib/logstash-core_jars.rb
@@ -9,7 +9,9 @@ rescue LoadError
   require 'com/fasterxml/jackson/core/jackson-annotations/2.9.1/jackson-annotations-2.9.1.jar'
   require 'org/apache/logging/log4j/log4j-slf4j-impl/2.6.2/log4j-slf4j-impl-2.6.2.jar'
   require 'com/fasterxml/jackson/dataformat/jackson-dataformat-cbor/2.9.1/jackson-dataformat-cbor-2.9.1.jar'
+  require 'org/codehaus/janino/commons-compiler/3.0.7/commons-compiler-3.0.7.jar'
   require 'com/fasterxml/jackson/core/jackson-core/2.9.1/jackson-core-2.9.1.jar'
+  require 'org/codehaus/janino/janino/3.0.7/janino-3.0.7.jar'
 end
 
 if defined? Jars
@@ -20,5 +22,7 @@ if defined? Jars
   require_jar( 'com.fasterxml.jackson.core', 'jackson-annotations', '2.9.1' )
   require_jar( 'org.apache.logging.log4j', 'log4j-slf4j-impl', '2.6.2' )
   require_jar( 'com.fasterxml.jackson.dataformat', 'jackson-dataformat-cbor', '2.9.1' )
+  require_jar( 'org.codehaus.janino', 'commons-compiler', '3.0.7' )
   require_jar( 'com.fasterxml.jackson.core', 'jackson-core', '2.9.1' )
+  require_jar( 'org.codehaus.janino', 'janino', '3.0.7' )
 end