… logstash-core/gemspec_jars.rb
… logstash-core/src/main/java/org/logstash/config/ir/compiler/DatasetCompiler.java