@original-brownbear 2017-10-19
@andrewvc there we go, this should make the Java exec a lot more interesting for people to try out :)