@andrewvc 2017-10-21
Very impressive work @original-brownbear !

Reformatting your Apache results for clarity for me (+ others):
```
bin/logstash -w 4 -b 128 -f ~/tmp/logstash.cfg --experimental-java-execution  
1080.48s user 25.05s system 430% cpu 4:16.59 total
bin/logstash -w 4 -b 128 -f ~/tmp/logstash.cfg  
1202.31s user 24.85s system 428% cpu 4:46.39 total

User CPU Reduction: 11%
System CPU: 0%
Wallclock: 11% reduction
```

It's always nice when userspace CPU correlates exactly to wallclock since users don't have to do any extra work to reap that practical benefit!