@original-brownbear 2017-10-24
no changes in #83f407, just rebased to fix conflicts