@original-brownbear 2017-10-25
@jakelandis reassigning you by orders from @andrewvc . Ping me on Slack if you want a 1:1 intro for this whole topic :)