@jakelandis 2017-10-26
@original-brownbear - Do you have an example of the generated source for a basic pipeline ?