@original-brownbear 2017-10-27
@jakelandis sure, so for a simple pipeline output snippet (not we only code generate the output part, for now, full stage including filters would be the next step :)) `output { stdout{} }`
You get this (without the comments I added :P):

```java
class CompiledDataset3 {
    // multi_receive method call site extracted from the OutputDelegator instance
    private final org.jruby.internal.runtime.methods.MixedModeIRMethod field0;
    // arguments array for calling that callsite, since it requires its args to be wrapped in an array, 
    //  the Callsite API doesn't take varargs 
    private final org.jruby.runtime.builtin.IRubyObject[] field1;
    // The OutputDelegator instance
    private final org.jruby.RubyObjectVar9 field2;

    public CompiledDataset3(Object field0, Object field1, Object field2) {
        this.field0 = (org.jruby.internal.runtime.methods.MixedModeIRMethod) field0;
        this.field1 = (org.jruby.runtime.builtin.IRubyObject[]) field1;
        this.field2 = (org.jruby.RubyObjectVar9) field2;
    }

    public Collection compute(RubyArray batch, boolean flush, boolean shutdown) {
        field1[0] = batch;
        field0.call(RubyUtil.RUBY.getCurrentContext(), field2, RubyUtil.LOGSTASH_MODULE,
            "multi_receive", field1, Block.NULL_BLOCK
        );
        return null;
    }

    public void clear() {
    }
}
```