@jakelandis 2017-10-27
I scanned the code and admittedly don't fully understand the changes presented here....however, I don't think at this stage it is necessary. 

* the changes are behind a feature flag
* this is part of a much larger change set 
* there may be refactoring of this in newer iterations
* POC/WIP to integrate early shouldn't need line by line reviews. 

In general this design and direction is really cool, and am in full support.  As briefly discussed in Slack, we should create review milestones once the final iterations of the whole picture are landing. 

In the meantime LGTM, lets gets this merged to master.